deleted a few files, renaming components, reorganizing
