+++ client/components/columnNode.jsx
@@ -2,9 +2,6 @@ import React, { memo, useContext } from 'react';
 import { Handle } from 'react-flow-renderer';
 
 import { DiagramContext } from '../state/contexts';
-// import TestNodeRow from './testNodeRow';
-import TableContents from './tableContents';
-import TableContainer from './tableContainer';
 
 // testing to see if tableFlowNode can render testNode
